+++ packages/vuikit/src/modal/__dev__/stories.js
@@ -2,3 +2,4 @@ import { storiesOf } from '@storybook/vue'
 
 storiesOf('Modal', module)
   .add('Default', () => require('./default.vue').default)
+  .add('Full', () => require('./full.vue').default)