+++ packages/vuikit/src/modal/components/index.js
@@ -1,6 +1,4 @@
 export { default as Modal } from './modal'
 export { default as ModalFull } from './modal-full'
-
-export { ElModalClose as ModalClose } from '../elements'
-export { ElModalFullClose as ModalFullClose } from '../elements'
 export { ElModalTitle as ModalTitle } from '../elements'
+export { ElModalClose as ModalClose } from '../elements'