+++ packages/vuikit/src/modal/components/modal-full.js
@@ -1,19 +1,34 @@
 import core from './core'
 import Transition from '../transition'
-import { TOGGLE } from '../constants'
-import { ElModalFull, ElModalDialog } from '../elements'
+import { ElModalFull } from '../elements'
 
-import { each, assign } from '@vuikit/utils/lang'
+import { each } from '@vuikit/utils/lang'
 import VkHeightViewport from 'vuikit/src/height-viewport'
 
 export default {
   name: 'VkModalFull',
   extends: core,
+  props: {
+    closeBtn: {
+      type: Boolean,
+      default: false
+    },
+    closeBtnLarge: {
+      type: Boolean,
+      default: false
+    }
+  },
   directives: {
     VkHeightViewport
   },
   render (h) {
-    const def = {
+    Object.keys(this.$slots).forEach(slot => each(this.$slots[slot], node => {
+      if (node.fnOptions && node.fnOptions.name === 'ElModalClose') {
+        node.data.staticClass = 'uk-modal-close-full'
+      }
+    }))
+
+    const modal = h(ElModalFull, {
       props: {
         expand: 'full'
       },
@@ -21,23 +36,8 @@ export default {
         name: 'show',
         value: this.show
       }]
-    }
-
-    Object.keys(this.$slots).forEach(slot => each(this.$slots[slot], node => {
-      if (node.fnOptions && node.fnOptions.name === 'VkModalFullClose') {
-        assign(node.data, {
-          on: { click: e => this.$emit(TOGGLE, false) }
-        })
-      }
-    }))
-
-    const modal = h(ElModalFull, def, [
-      h(ElModalDialog, {
-        class: 'uk-flex uk-flex-center uk-flex-middle',
-        directives: [{
-          name: 'vk-height-viewport'
-        }]
-      }, this.$slots.default)
+    }, [
+      this.$slots.default
     ])
 
     return h(Transition, [ modal ])