+++ packages/vuikit/src/modal/elements/modal--dialog.js
@@ -2,6 +2,7 @@ import { mergeData } from '@vuikit/utils/vue'
 
 export default {
   functional: true,
+  name: 'ElModalDialog',
   render (h, { props, data, children }) {
     return h('div', mergeData(data, {
       class: 'uk-modal-dialog'