+++ packages/vuikit/src/modal/elements/modal-full.js
@@ -1,13 +1,24 @@
+import ElModalDialog from './modal--dialog'
 import { mergeData } from '@vuikit/utils/vue'
 
 export default {
   functional: true,
-  render (h, { children, data, props }) {
+  name: 'ElModalFull',
+  render (h, { data, props, children }) {
     return h('div', mergeData(data, {
       class: 'uk-modal uk-modal-full',
       style: {
         display: 'block'
       }
-    }), children)
+    }), [
+
+      h(ElModalDialog, {
+        class: 'uk-flex uk-flex-center uk-flex-middle',
+        directives: [{
+          name: 'vk-height-viewport'
+        }]
+      }, children)
+
+    ])
   }
 }