+++ packages/vuikit/src/modal/elements/modal.js
@@ -1,23 +1,40 @@
+import ElModalDialog from './modal--dialog'
 import { mergeData } from '@vuikit/utils/vue'
 
 export default {
   functional: true,
+  name: 'ElModal',
   props: {
-    expand: {
+    // vertically expands the modal
+    expanded: {
+      type: Boolean,
+      default: false
+    },
+    // vertically centers the modal
+    centered: {
       type: Boolean,
       default: false
     }
   },
-  render (h, { children, data, props }) {
-    const { expand } = props
+  render (h, { data, props, children }) {
+    const { expanded, centered } = props
 
     return h('div', mergeData(data, {
       class: ['uk-modal', {
-        'uk-modal-container': expand
+        'uk-modal-container': expanded,
+        'uk-flex uk-flex-top': centered
       }],
       style: {
-        display: 'block'
+        display: centered ? 'flex' : 'block'
       }
-    }), children)
+    }), [
+
+      h(ElModalDialog, {
+        class: {
+          'uk-margin-auto-vertical': centered
+        }
+      }, children)
+
+    ])
   }
 }